Update to be usable on SQL
- asked TA for help for this...
- added more pokemon
- added prof motherfucking oak

+++ PokeDB.sql
@@ -1,10 +1,11 @@
---Database Table Creation
-drop table Creates;
-drop table Modifies;
-drop table Account;
-drop table DBManager;
-drop table Trainer;
-drop table Pokemon;
+set Foreign_key_checks=0;
+
+drop table Creates if exists enrolled cascade;
+drop table Modifies if exists enrolled cascade;
+drop table Account if exists enrolled cascade;
+drop table DBManager if exists enrolled cascade;
+drop table Trainer if exists enrolled cascade;
+drop table Pokemon if exists enrolled cascade;
 
 CREATE TABLE Pokemon
 (Pokemon_ID integer not null PRIMARY KEY,
@@ -16,7 +17,7 @@ CREATE TABLE Trainer
 (trainer_ID integer not null PRIMARY KEY,
 TName varCHAR(20),
 TGender varCHAR(6),
-THometown varCHAR(30),
+THometown varCHAR(40),
 TWin integer,
 TLoss integer,
 check (trainer_ID >= 0 AND TGender in ('Male', 'Female') AND TWin >=0 AND TLoss >=0));
@@ -48,19 +49,55 @@ insert into Pokemon values
 insert into Pokemon values
 (00000003,'Venusaur', 1);
 
+insert into Pokemon values
+(00000004, 'Charmander', 1);
+
+insert into Pokemon values
+(00000005, 'Charmeleon', 1);
+
+insert into Pokemon values
+(00000006, 'Charizard', 1);
+
+insert into Pokemon values
+(00000007, 'Squirtle', 1);
+
+insert into Pokemon values
+(00000008, 'Wartortle', 1);
+
+insert into Pokemon values
+(00000009, 'Blastoise', 1);
+
+insert into Pokemon values
+(00000095, 'Onix', 1);
+
 insert into Trainer values
 (00000001, 'Red', 'Male', 'Pallet Town', 100, 0);
 
 insert into Trainer values
-(00000002, 'Brock', 'Male', 'Pewter City', 100, 0);
+(00000002, 'Blue', 'Male', 'Pallet Town', 99, 1);
+
+insert into Trainer values
+(00000003, 'Frigo Oak', 'Male', 'Pallet Town', 0, 0);
+
+insert into Trainer values
+(00000004, 'Brock', 'Male', 'Pewter City', 1, 0);
 
 insert into TrainedPokemon values
-(00000001,'Bulbasaur', 'Red', 00000001);
+(00000004,'Charmander', 'Red', 00000001);
+
+insert into TrainedPokemon values
+(00000007,'Squirtle', 'Blue', 00000002);
+
+insert into TrainedPokemon values
+(00000001,'Bulbasaur', 'Frigo Oak', 00000003);
+
+insert into TrainedPokemon values
+(00000095,'Onix', 'Brock', 00000004);
 
 insert into GymLeader values
-(00000002, 'Pewter Gym', 'Boulder Badge');
+(00000003, 'Pewter Gym', 'Boulder Badge');
 
 insert into DBManager values
-(00000001);
+(00000003);
 
-commit;
+set Foreign_key_checks=1;
